Handle build bake through bake, bake-files and bake-targets
