+++ test/Dockerfile-bake
@@ -1,0 +1,4 @@
+FROM alpine
+
+ARG name=world
+RUN echo "Hello ${name}!"
